backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
google-generativeai
pypdf
python-dotenv
numpy
pydantic
langchain==0.2.0
langchain-community==0.2.0
langchain-core
langchain-text-splitters
langchain-chroma
langchain-google-genai
watchdog
sentence-transformers
langchain-huggingface
